[v4.5.1] Harden backend consistency and env naming hygiene
