+++ README.md
@@ -9,10 +9,10 @@ A Next.js application featuring AI-powered skin analysis solutions and evidence-
 - High-precision region detection and analysis
 - SDK/API integration for e-commerce and retail
 
-### 🤖 CrystalAI Recommendation System
+### 🤖 Recommendation Engine
 - **Interactive Demo**: `/ja/recommendation-demo` or `/en/recommendation-demo`
 - Real-time product recommendations based on skin type and condition
-- CrystalAI 82-model ensemble integration
+- Recommendation engine API integration
 - Rate-limited demo access (3 requests per 10 minutes)
 - Analytics integration with GA4
 - Human-in-the-loop review queue system
@@ -81,13 +81,32 @@ npm run test-email
 # Required for email functionality
 NEXT_PUBLIC_SITE_URL=your-site-url
 NEXT_PUBLIC_GA_ID=your-ga-id
-GMAIL_USER=your-email@gmail.com
-GMAIL_APP_PASSWORD=your-app-password
+EMAIL_USER=your-email@gmail.com
+EMAIL_APP_PASSWORD=your-app-password
+EMAIL_FROM_NAME=Visage AI Consulting
+ADMIN_EMAIL=admin@visageaiconsulting.com
 
-# Required for CrystalAI features
-NEXT_PUBLIC_FASTAPI_URL=https://visage-ai-api.vercel.app
-NEXT_PUBLIC_FASTAPI_TOKEN=your-api-token
+# Required for recommendation API features
+NEXT_PUBLIC_API_BASE_URL=https://api.visageaiconsulting.com
+API_BASE_URL=https://api.visageaiconsulting.com
+API_AUTH_TOKEN=your-api-token
 ```
+
+### Compliance Hygiene Check (Required Before Merge)
+Run the guardrail check before releasing:
+
+```bash
+npm run check:compliance-hygiene
+```
+
+This check fails if active source reintroduces:
+- legacy backend domain (`https://visage-ai-api.vercel.app`),
+- legacy route usage (`/api/v2/crystalai/*`) in webapp active paths,
+- sensitive `NEXT_PUBLIC_*` env naming patterns.
+
+### Legacy Compatibility Notes
+- Canonical recommendation routes are `/api/v2/recommendations`, `/api/v2/demo`, and `/api/v2/demo/rate-limit`.
+- Backend may retain hidden legacy aliases for compatibility (`/api/v2/crystalai/*`), but webapp callers must use canonical routes.
 
 ## Learn More
