+++ app/[locale]/recommendation-demo/page.tsx
@@ -1,4 +1,4 @@
-import { CrystalAIRecommendationDemo } from "@/components/landing/CrystalAIRecommendationDemo";
+import { RecommendationDemo } from "@/components/landing/RecommendationDemo";
 import type { Metadata } from "next";
 
 export async function generateMetadata({
@@ -7,8 +7,8 @@ export async function generateMetadata({
   params: { locale: string };
 }): Promise<Metadata> {
   const title = locale === 'ja'
-    ? 'CrystalAI レコメンデーション - Visage AI'
-    : 'CrystalAI Recommendations - Visage AI';
+    ? 'おすすめデモ - Visage AI'
+    : 'Recommendation Demo - Visage AI';
   const description = locale === 'ja'
     ? 'リアルタイムでスキンケア製品レコメンデーションを試す'
     : 'Try real-time skincare product recommendations';
@@ -27,7 +27,6 @@ export default function RecommendationDemoPage({
   params: { locale: string };
 }) {
   return (
-    <CrystalAIRecommendationDemo locale={locale as 'ja' | 'en'} />
+    <RecommendationDemo locale={locale as 'ja' | 'en'} />
   );
 }
-
