+++ public/compliance/privacy-fallback-en.html
@@ -1,0 +1,32 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Visage AI Privacy Notice (Fallback)</title>
+    <style>
+      body { font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif; margin: 0; padding: 24px; line-height: 1.6; color: #111827; }
+      .wrap { max-width: 860px; margin: 0 auto; }
+      h1 { margin-bottom: 8px; }
+      .badge { display: inline-block; background: #eff6ff; color: #1d4ed8; padding: 4px 10px; border-radius: 999px; font-size: 12px; margin-bottom: 16px; }
+      .card { background: #f8fafc; border: 1px solid #e2e8f0; border-radius: 12px; padding: 16px; margin: 12px 0; }
+    </style>
+  </head>
+  <body>
+    <main class="wrap">
+      <h1>Visage AI Privacy Notice (Fallback Copy)</h1>
+      <div class="badge">Emergency backup content</div>
+      <p>This page is a static fallback for temporary publication continuity.</p>
+      <div class="card">
+        <strong>On-device processing:</strong> Core skin analysis and Deep Skin visualization are processed on-device (Core Image, Metal, Vision).
+      </div>
+      <div class="card">
+        <strong>First-party backend:</strong> Data is sent only to <code>api.visageaiconsulting.com</code>, operated by Faceless (Visage AI).
+      </div>
+      <div class="card">
+        <strong>No third-party AI transmission:</strong> We do not send user data to OpenAI, Anthropic, Gemini, or other third-party AI providers.
+      </div>
+      <p>Primary policy URL: <a href="https://www.visageaiconsulting.com/en/privacy">https://www.visageaiconsulting.com/en/privacy</a></p>
+    </main>
+  </body>
+</html>
