+++ public/compliance/privacy-fallback-ja.html
@@ -1,0 +1,32 @@
+<!doctype html>
+<html lang="ja">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Visage AI プライバシー通知（フォールバック）</title>
+    <style>
+      body { font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif; margin: 0; padding: 24px; line-height: 1.7; color: #111827; }
+      .wrap { max-width: 860px; margin: 0 auto; }
+      h1 { margin-bottom: 8px; }
+      .badge { display: inline-block; background: #eff6ff; color: #1d4ed8; padding: 4px 10px; border-radius: 999px; font-size: 12px; margin-bottom: 16px; }
+      .card { background: #f8fafc; border: 1px solid #e2e8f0; border-radius: 12px; padding: 16px; margin: 12px 0; }
+    </style>
+  </head>
+  <body>
+    <main class="wrap">
+      <h1>Visage AI プライバシー通知（フォールバック版）</h1>
+      <div class="badge">緊急時バックアップ用コンテンツ</div>
+      <p>このページは、公開継続性確保のための静的フォールバックです。</p>
+      <div class="card">
+        <strong>端末内処理:</strong> 肌解析および Deep Skin 可視化は端末内（Core Image / Metal / Vision）で処理されます。
+      </div>
+      <div class="card">
+        <strong>自社バックエンド:</strong> 送信先は Faceless（Visage AI）が運営する <code>api.visageaiconsulting.com</code> のみです。
+      </div>
+      <div class="card">
+        <strong>第三者AI送信なし:</strong> OpenAI、Anthropic、Gemini 等の第三者AIサービスへユーザーデータを送信しません。
+      </div>
+      <p>正規ポリシーURL: <a href="https://www.visageaiconsulting.com/ja/privacy">https://www.visageaiconsulting.com/ja/privacy</a></p>
+    </main>
+  </body>
+</html>
